+++ include/datastructure/Timetable.hpp
@@ -148,7 +148,7 @@ public:
   };
 
   [[nodiscard]] int                 max_t() const;
-  [[nodiscard]] std::pair<int, int> time_interval(int train_index) const;
+  [[nodiscard]] std::pair<int, int> time_interval(size_t train_index) const;
   [[nodiscard]] std::pair<int, int>
   time_interval(const std::string& train_name) const {
     return time_interval(train_list.get_train_index(train_name));
